+++ spicerack/tests/unit/test_netbox.py
@@ -6,7 +6,7 @@ import pynetbox
 import pytest
 import yaml
 
-from spicerack.netbox import Netbox, NetboxAPIError, NetboxError
+from spicerack.netbox import Netbox, NetboxAPIError, NetboxError, NetboxHostNotFoundError
 from spicerack.tests import get_fixture_path
 
 
@@ -26,6 +26,17 @@ def _fake_host():
     fake_host = mock.Mock()
     fake_host.name = 'test'
     fake_host.status = 'Active'
+    fake_host.serialize = mock.Mock(return_value={'name': 'test'})
+    fake_host.save = mock.Mock(return_value=True)
+    return fake_host
+
+
+def _fake_virtual_host():
+    """Return a mocked Netbox host."""
+    fake_host = mock.Mock()
+    fake_host.name = 'test'
+    fake_host.status = 'Active'
+    fake_host.cluster.name = 'testcluster'
     fake_host.serialize = mock.Mock(return_value={'name': 'test'})
     fake_host.save = mock.Mock(return_value=True)
     return fake_host
@@ -59,9 +70,10 @@ def test_netbox_choices_api_devices_missing(mocked_pynetbox):
 def test_netbox_fetch_host_status_nohost(mocked_pynetbox):
     """Test the error scenario where the host is not found."""
     mocked_pynetbox().dcim.devices.get.return_value = None
+    mocked_pynetbox().virtualization.virtual_machines.get.return_value = None
     mocked_pynetbox().dcim.choices = _get_choices_mock()
     netbox = Netbox(NETBOX_URL, NETBOX_TOKEN, dry_run=True)
-    with pytest.raises(NetboxError, match='host not found'):
+    with pytest.raises(NetboxHostNotFoundError):
         netbox.fetch_host_status('host')
 
 
@@ -69,6 +81,7 @@ def test_netbox_fetch_host_status_nohost(mocked_pynetbox):
 def test_netbox_fetch_host_status_error(mocked_pynetbox):
     """Test the error scenario where the host is not found."""
     mocked_pynetbox().dcim.devices.get = mock.Mock(side_effect=_request_error())
+    mocked_pynetbox().virtualization.virtual_machines.get = mock.Mock(side_effect=_request_error())
     mocked_pynetbox().dcim.choices = _get_choices_mock()
     netbox = Netbox(NETBOX_URL, NETBOX_TOKEN, dry_run=True)
     with pytest.raises(NetboxError, match='error retrieving host'):
@@ -79,6 +92,7 @@ def test_netbox_fetch_host_status_error(mocked_pynetbox):
 def test_fetch_host_status(mocked_pynetbox):
     """Test fetching host status."""
     mocked_pynetbox().dcim.devices.get.return_value = _fake_host()
+    mocked_pynetbox().virtualization.virtual_machines.get.return_value = None
     mocked_pynetbox().dcim.choices = _get_choices_mock()
     netbox = Netbox(NETBOX_URL, NETBOX_TOKEN, dry_run=True)
 
@@ -102,6 +116,7 @@ def test_put_host_status_good_status(mocked_pynetbox):
     """Test setting a status and it working."""
     fake_host = _fake_host()
     mocked_pynetbox().dcim.devices.get.return_value = fake_host
+    mocked_pynetbox().virtualization.virtual_machines.get.return_value = None
     mocked_pynetbox().dcim.choices = _get_choices_mock()
     netbox = Netbox(NETBOX_URL, NETBOX_TOKEN, dry_run=False)
 
@@ -115,6 +130,7 @@ def test_put_host_status_save_failure(mocked_pynetbox):
     """Test save failure."""
     fake_host = _fake_host()
     mocked_pynetbox().dcim.devices.get.return_value = fake_host
+    mocked_pynetbox().virtualization.virtual_machines.get.return_value = None
     mocked_pynetbox().dcim.choices = _get_choices_mock()
     netbox = Netbox(NETBOX_URL, NETBOX_TOKEN, dry_run=False)
     fake_host.save.return_value = False
@@ -140,6 +156,7 @@ def test_put_host_status_error(mocked_pynetbox):
     """Test exception during save."""
     fake_host = _fake_host()
     mocked_pynetbox().dcim.devices.get.return_value = fake_host
+    mocked_pynetbox().virtualization.virtual_machines.get.return_value = None
     mocked_pynetbox().dcim.choices = _get_choices_mock()
     netbox = Netbox(NETBOX_URL, NETBOX_TOKEN, dry_run=False)
     fake_host.save.side_effect = _request_error()
@@ -153,7 +170,61 @@ def test_fetch_host_detail(mocked_pynetbox):
     """Test fetching host detail."""
     fake_host = _fake_host()
     mocked_pynetbox().dcim.devices.get.return_value = fake_host
+    mocked_pynetbox().virtualization.virtual_machines.get.return_value = None
     mocked_pynetbox().dcim.choices = _get_choices_mock()
     netbox = Netbox(NETBOX_URL, NETBOX_TOKEN, dry_run=True)
-    netbox.fetch_host_detail('test')
+    detail = netbox.fetch_host_detail('test')
     assert fake_host.serialize.called
+    assert not detail['is_virtual']
+
+
+@mock.patch('pynetbox.api')
+def test_fetch_host_detail_vm(mocked_pynetbox):
+    """Virtual machines should have is_virtual == True and a cluster name."""
+    fake_host = _fake_virtual_host()
+    mocked_pynetbox().dcim.devices.get.return_value = None
+    mocked_pynetbox().dcim.choices = _get_choices_mock()  # this is auto populated so must work
+    mocked_pynetbox().virtualization.virtual_machines.get.return_value = fake_host
+    netbox = Netbox(NETBOX_URL, NETBOX_TOKEN, dry_run=True)
+    detail = netbox.fetch_host_detail('test')
+    assert fake_host.serialize.called
+    assert detail['is_virtual']
+    assert detail['cluster_name'] == 'testcluster'
+
+
+@mock.patch('pynetbox.api')
+def test_fetch_host_status_vm(mocked_pynetbox):
+    """Virtual machines should return status just like devices."""
+    mocked_pynetbox().dcim.devices.get.return_value = None
+    mocked_pynetbox().virtualization.virtual_machines.get.return_value = _fake_virtual_host()
+    mocked_pynetbox().dcim.choices = _get_choices_mock()
+    netbox = Netbox(NETBOX_URL, NETBOX_TOKEN, dry_run=True)
+
+    assert netbox.fetch_host_status('host') == 'Active'
+
+
+@mock.patch('pynetbox.api')
+def test_fetch_host_status_vm_error(mocked_pynetbox):
+    """Virtual machines should raise an exception on an API error."""
+    mocked_pynetbox().dcim.devices.get.return_value = None
+    mocked_pynetbox().virtualization.virtual_machines.get = mock.Mock(side_effect=_request_error())
+    mocked_pynetbox().dcim.choices = _get_choices_mock()
+    netbox = Netbox(NETBOX_URL, NETBOX_TOKEN, dry_run=True)
+
+    with pytest.raises(NetboxAPIError, match="error retrieving VM"):
+        netbox.fetch_host_status('host')
+
+
+@mock.patch('pynetbox.api')
+def test_set_host_status_vm(mocked_pynetbox):
+    """Virtual machines should raise an exception if you try to set the status."""
+    fake_vm = _fake_virtual_host()
+    mocked_pynetbox().dcim.devices.get.return_value = None
+    mocked_pynetbox().virtualization.virtual_machines.get.return_value = fake_vm
+    mocked_pynetbox().dcim.choices = _get_choices_mock()
+    netbox = Netbox(NETBOX_URL, NETBOX_TOKEN, dry_run=True)
+
+    with pytest.raises(NetboxHostNotFoundError):
+        netbox.put_host_status('test', 'Active')
+
+    assert not fake_vm.save.called